refactor: request changes
- relocated SearchField to dedicated file
- worked on event resolution formating
