+++ flexmeasures/api/common/schemas/generic_assets.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
-from shlex import join, split
-
 from flask import abort
-from marshmallow import fields, ValidationError
+from marshmallow import fields
 from sqlalchemy import select
 
 from flexmeasures.data import db
@@ -25,17 +23,3 @@ class AssetIdField(fields.Integer):
 
     def _serialize(self, asset: GenericAsset, attr, data, **kwargs) -> int:
         return asset.id
-
-
-class SearchFilterField(fields.Str):
-    """Field that represents a search filter."""
-
-    def _deserialize(self, value, attr, data, **kwargs) -> list[str]:
-        try:
-            search_terms = split(value)
-        except ValueError as e:
-            raise ValidationError(str(e))
-        return search_terms
-
-    def _serialize(self, value: list[str], attr, obj, **kwargs) -> str:
-        return join(value)
